back_end/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.8
djangorestframework==3.15.2
idna==3.7
pillow==10.4.0
psycopg==3.2.1
psycopg-binary==3.2.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
